fix(ci,groove,docs): harden recorder build and 1.1 manual
- fix Groove Recorder signed-offset conversion for strict Clang builds
- add symmetric maximum early/late recorder regression coverage
- upgrade GitHub Actions cache usage to Node 24-capable actions/cache v6
- upgrade release artifact upload to actions/upload-artifact v7
- add workflow regression checks against deprecated Node 20 action majors
- add semantic manual chapter navigation and contents overview
- document Pre-Count behavior, settings and external-sync semantics
- add Swing, Groove and Humanize comparison and interaction guidance
- add a guided live Groove recording workflow
- distinguish global Pre-Count from Recorder Count-In
- fix Groove screenshot anchoring and prevent text overlap
- correct Eurorack power-connector and boot-safety documentation
- add Pre-Count and AUTHOR popover screenshots
- clarify Easter-egg intro-to-game controls
- harden manual validation against obsolete hardware wording and unsafe image anchors

+++ src/services/groove_recorder.cpp
@@ -153,9 +153,12 @@ bool GrooveRecorder::capture(
     const std::uint64_t boundedMagnitude = std::min<std::uint64_t>(
         scaled,
         static_cast<std::uint64_t>(kCustomGrooveMaximumOffset256));
-    const std::int16_t signedOffset = late
-        ? static_cast<std::int16_t>(boundedMagnitude)
-        : -static_cast<std::int16_t>(boundedMagnitude);
+    // Negating an int16_t promotes it to int. Keep the sign operation in a
+    // type that can represent the complete intermediate range explicitly so
+    // Clang -Wconversion/-Werror builds do not depend on an implicit narrowing.
+    const std::int32_t signedOffset = late
+        ? static_cast<std::int32_t>(boundedMagnitude)
+        : -static_cast<std::int32_t>(boundedMagnitude);
     pattern.offsets256[step] = static_cast<std::int8_t>(signedOffset);
     capturedMask_ |= (1ULL << step);
     playheadStep_ = step;
